requirements.txt
streamlit 
ultralytics 
opencv-python 
Pillow 
torch 
torchvision